requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.0
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
filelock==3.16.1
    # via
    #   huggingface-hub
    #   transformers
fsspec==2024.10.0
    # via huggingface-hub
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
huggingface-hub==0.26.5
    # via
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marqo==3.9.2
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.0
    # via transformers
packaging==24.2
    # via
    #   huggingface-hub
    #   marqo
    #   transformers
pydantic==2.9.2
    # via
    #   fastapi
    #   marqo
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.12
    # via fastapi
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   marqo
    #   transformers
rich==13.9.2
    # via typer
safetensors==0.4.5
    # via transformers
semantic-text-splitter==0.19.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.6
    # via fastapi
tokenizers==0.21.0
    # via transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.47.0
    # via -r requirements.in
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   typer
urllib3==2.2.3
    # via
    #   marqo
    #   requests
uvicorn[standard]==0.31.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
